src/model/requirements.txt
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.10.0
filelock==3.12.0
fonttools==4.39.4
idna==2.10
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psutil==5.9.5
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-toolbelt==1.0.0
roboflow==1.0.9
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.23.1
six==1.16.0
sympy==1.12
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.105
urllib3==1.26.15
wget==3.2
